+++ compose.yml
@@ -28,10 +28,12 @@ services:
       retries: 5
 
   server:
+    env_file:
+      - .env
     build:
       context: ./
     environment:
-      NODE_ENV: production
+      MISTRAL_API_KEY: "${MISTRAL_API_KEY}"
       DATABASE_URL: "postgresql://${username}:${password}@db:5432/${database}?schema=public"
     ports:
       - 3000:3000
